Add documentation for how to work with a non-default codebase
Add `--working-copy` option to drush make to download projects with
version control. This is necessary if you as a developer want to work
with the code. It is also a prerequisite for switching to another
branch og repository.

Add instructions for checking out a branch in a fork of the original
repository.